rough draft

+++ test/test.cpp
@@ -4,7 +4,7 @@
 
 using namespace std;
 int main(int argc, char** argv){
-  int  n     = argc>1 ? atoi(argv[1]) : 2;
+  double  n     = argc>1 ? atof(argv[1]) : 2.0;
   double reltol=1e-3;
   auto f =[n](double x) -> double { return pow(x,n); }; 
   auto intf  = gauss_kronrod(f,reltol);
